+++ ecs.php
@@ -1,4 +1,5 @@
 <?php declare(strict_types=1);
+
 /*
  * (c) shopware AG <info@shopware.com>
  * For the full copyright and license information, please view the LICENSE
@@ -10,29 +11,27 @@ use PhpCsFixer\Fixer\Comment\HeaderCommentFixer;
 use PhpCsFixer\Fixer\FunctionNotation\NativeFunctionInvocationFixer;
 use Symplify\EasyCodingStandard\Config\ECSConfig;
 
-return static function (ECSConfig $containerConfigurator): void {
-    $services = $containerConfigurator->services();
-
-    $services->set(HeaderCommentFixer::class)
-        ->call('configure', [['header' => '(c) shopware AG <info@shopware.com>
+return static function (ECSConfig $ecsConfig): void {
+    $ecsConfig->ruleWithConfiguration(HeaderCommentFixer::class, [
+        'header' => '(c) shopware AG <info@shopware.com>
 For the full copyright and license information, please view the LICENSE
-file that was distributed with this source code.', 'separate' => 'bottom', 'location' => 'after_declare_strict', 'comment_type' => 'comment']]);
+file that was distributed with this source code.',
+        'separate' => 'bottom',
+        'location' => 'after_declare_strict',
+        'comment_type' => 'comment'
+    ]);
 
-    $services->set(NativeFunctionInvocationFixer::class)
-        ->call('configure', [[
-            'include' => [NativeFunctionInvocationFixer::SET_ALL],
-            'scope' => 'namespaced',
-        ]]);
+    $ecsConfig->ruleWithConfiguration(NativeFunctionInvocationFixer::class, [
+        'include' => [NativeFunctionInvocationFixer::SET_ALL],
+        'scope' => 'namespaced',
+    ]);
 
-    $services->set(MbStrFunctionsFixer::class);
+    $ecsConfig->rule(MbStrFunctionsFixer::class);
 
-    $parameters = $containerConfigurator->parameters();
+    $ecsConfig->cacheDirectory(__DIR__ . '/var/cache/cs_fixer');
+    $ecsConfig->cacheNamespace('SwagMigrationAssistant');
 
-    $parameters->set('cache_directory', __DIR__ . '/var/cache/cs_fixer');
-
-    $parameters->set('cache_namespace', 'SwagMigrationAssistant');
-
-    $parameters->set('paths', [
+    $ecsConfig->paths([
         __DIR__ . '/Command',
         __DIR__ . '/Controller',
         __DIR__ . '/Core',